requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.5
django-debug-toolbar==3.2
django-taggit==1.3.0
googletrans==3.0.0
gTTS==2.2.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==3.4.0
Markdown==3.3.3
Pillow==8.1.0
psycopg2==2.8.6
pytz==2020.5
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
